+++ packages/brick_offline_first_build/test/offline_first_generator/test_futures.dart
@@ -108,9 +108,6 @@ class FuturesAdapter extends OfflineFirstAdapter<Futures> {
   FuturesAdapter();
 
   @override
-  final fieldsToOfflineFirstRuntimeDefinition =
-      <String, RuntimeOfflineFirstDefinition>{};
-  @override
   final Map<String, RuntimeSqliteColumnDefinition> fieldsToSqliteColumns = {
     'primaryKey': const RuntimeSqliteColumnDefinition(
       association: false,